# backend/server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CloudHub",
    "description": "API endpoints for CloudHub",
    "version": "1.0.0"
  },
  "schemes": ["http"],
  "basePath": "/cloudhub/v1",
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/cloudhub/v1/config": {
      "get": {
        "tags": ["config"],
        "summary": "Returns the global application configuration",
        "description": "All global application configurations",
        "responses": {
          "200": {
            "description": "Returns an object with the global configurations",
            "schema": {
              "$ref": "#/definitions/Config"
            }
          },
          "404": {
            "description": "Could not find global application config",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/cloudhub/v1/config/auth": {
      "get": {
        "tags": ["config"],
        "summary": "Returns the global application configuration for auth",
        "description": "All global application configuration for auth",
        "responses": {
          "200": {
            "description": "Returns an object with the global application configuration for auth",
            "schema": {
              "$ref": "#/definitions/AuthConfig"
            }
          },
          "404": {
            "description": "Could not find auth configuration",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": ["config"],
        "summary": "Updates the global application configuration for auth",
        "description": "Replaces the global application configuration for auth",
        "parameters": [
          {
            "name": "auth",
            "in": "body",
            "description": "Auth configuration update object",
            "schema": {
              "$ref": "#/definitions/AuthConfig"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object with the updated auth configuration",
            "schema": {
              "$ref": "#/definitions/AuthConfig"
            }
          },
          "404": {
            "description": "Could not find auth configuration",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/cloudhub/v1/org_config": {
      "get": {
        "tags": ["organization config"],
        "summary": "Retrieve the organization configuration",
        "description": "Organization-specific configurations such as log viewer configs",
        "responses": {
          "200": {
            "description": "Returns an object with the organization-specific configurations",
            "schema": {
              "$ref": "#/definitions/OrganizationConfig"
            }
          },
          "404": {
            "description": "Could not find organization config",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/cloudhub/v1/org_config/logviewer": {
      "get": {
        "tags": ["organization config"],
        "summary": "Retrieve the organization-specific log viewer configurations",
        "description": "Retrieve the log viewer configurations for the user's current organization",
        "responses": {
          "200": {
            "description": "Returns an log viewer configuration object",
            "schema": {
              "$ref": "#/definitions/LogViewerConfig"
            }
          },
          "404": {
            "description": "Could not find the log viewer configuration for this organization",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": ["organization config"],
        "summary": "Update the log viewer configuration",
        "description": "Update the log viewer configuration for a specific organization",
        "parameters": [
          {
            "name": "logViewer",
            "in": "body",
            "description": "Log Viewer configuration update object",
            "schema": {
              "$ref": "#/definitions/LogViewerConfig"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object with the updated log viewer configurations",
            "schema": {
              "$ref": "#/definitions/LogViewerConfig"
            }
          },
          "404": {
            "description": "Could not find log viewer configurations for the specified organization",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Config": {
      "description": "Global application configuration",
      "type": "object",
      "properties": {
        "auth": {
          "$ref": "#/definitions/AuthConfig"
        }
      },
      "example": {
        "auth": {
          "superAdminNewUsers": true
        }
      }
    },
    "AuthConfig": {
      "description": "Global application configuration for auth",
      "type": "object",
      "required": ["superAdminNewUsers"],
      "properties": {
        "superAdminNewUsers": {
          "type": "boolean",
          "default": true
        }
      },
      "example": {
        "superAdminNewUsers": true
      }
    },
    "OrganizationConfig": {
      "description": "Configurations for a specific organization",
      "type": "object",
      "required": ["logViewer"],
      "properties": {
        "organization": {
          "type": "string",
          "readOnly": true
        },
        "logViewer": {
          "$ref": "#/definitions/LogViewerConfig"
        }
      },
      "example": {
        "organization": "default",
        "logViewer": {
          "columns": [
            {
              "name": "severity",
              "position": 0,
              "encodings": [
                {
                  "type": "label",
                  "value": "icon"
                },
                {
                  "type": "label",
                  "value": "text"
                },
                {
                  "type": "visibility",
                  "value": "visible"
                },
                {
                  "type": "color",
                  "name": "ruby",
                  "value": "emergency"
                },
                {
                  "type": "color",
                  "name": "rainforest",
                  "value": "info"
                },
                {
                  "type": "displayName",
                  "value": "Log Severity!"
                }
              ]
            },
            {
              "name": "messages",
              "position": 1,
              "encodings": [
                {
                  "type": "visibility",
                  "value": "hidden"
                }
              ]
            }
          ]
        }
      }
    },
    "LogViewerConfig": {
      "description": "Contains the organization-specific configuration for the log viewer",
      "type": "object",
      "required": ["columns"],
      "properties": {
        "columns": {
          "description": "Defines the order, names, and visibility of columns in the log viewer table",
          "type": "array",
          "items": {
            "$ref": "#/definitions/LogViewerColumn"
          }
        }
      },
      "example": {
        "columns": [
          {
            "name": "severity",
            "position": 0,
            "encodings": [
              {
                "type": "label",
                "value": "icon"
              },
              {
                "type": "label",
                "value": "text"
              },
              {
                "type": "visibility",
                "value": "visible"
              },
              {
                "type": "color",
                "name": "ruby",
                "value": "emergency"
              },
              {
                "type": "color",
                "name": "rainforest",
                "value": "info"
              },
              {
                "type": "displayName",
                "value": "Log Severity!"
              }
            ]
          },
          {
            "name": "messages",
            "position": 1,
            "encodings": [
              {
                "type": "visibility",
                "value": "hidden"
              }
            ]
          }
        ]
      }
    },
    "LogViewerColumn": {
      "description": "Contains the organization-specific configuration for the log viewer",
      "type": "object",
      "required": ["name", "encodings", "position"],
      "properties": {
        "name": {
          "description": "Unique identifier name of the column",
          "type": "string"
        },
        "position": {
          "type": "integer",
          "format": "int32"
        },
        "encodings": {
          "description": "Composable encoding options for the column",
          "type": "array",
          "items": {
            "description": "Type and value and optional name of an encoding",
            "type": "object",
            "required": ["type", "value"],
            "properties": {
              "type": {
                "type": "string"
              },
              "value": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          }
        }
      },
      "example": {
        "name": "severity",
        "position": 0,
        "encodings": [
          {
            "type": "label",
            "value": "icon"
          },
          {
            "type": "label",
            "value": "text"
          },
          {
            "type": "visibility",
            "value": "visible"
          },
          {
            "type": "color",
            "name": "ruby",
            "value": "emergency"
          },
          {
            "type": "color",
            "name": "rainforest",
            "value": "info"
          },
          {
            "type": "displayName",
            "value": "Log Severity!"
          }
        ]
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
  }
}